convert all to ts + tests
