+++ my-app/package.json
@@ -20,6 +20,7 @@
     "typescript": "3.4.5"
   },
   "devDependencies": {
+    "@types/jest": "24.0.15",
     "ts-node-dev": "1.0.0-pre.40"
   },
   "scripts": {
